# add x86_64 inline asm to direct_threading and optimized_direct_threading