`cargo criterion` results on Intel 7600k @ 4.8Ghz:

`direct threading`  time: [14.375 ms 14.376 ms 14.378 ms]

`optimized direct threading` time: [10.148 ms 10.155 ms 10.163 ms]